+++ webapp/src/components/GlobalNavigation/index.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 import { getUserInfo } from 'service/auth';
 import LoginNav from './LoginNav';
 import NonLoginNav from './NonLoginNav';
+import * as S from './GlobalNavigation.style';
 
-import * as S from './style';
-
-export default function Navigation() {
+export default function GlobalNavigation() {
   const userInfo = getUserInfo(); // {id, profileImg, nickname}
   return <S.Container>{userInfo ? <LoginNav userInfo={userInfo} /> : <NonLoginNav />}</S.Container>;
 }
